Make user forum test pass
I suspect this page is incorrectly displaying the <h2> tags as text at
the moment. I've changed the test to not look into those tags. It may be
we don't want to check <h2> specifically and can get away with the
strategy presented here.

I am committing this seperately to draw attention to this change.